+++ src/common/comment.ts
@@ -38,13 +38,3 @@ export interface Comment {
 	graphNodeId: string;
 	reactions?: Reaction[];
 }
-
-export interface CommentHandler {
-	commentController?: vscode.CommentController;
-	startReview(thread: vscode.CommentThread): Promise<void>;
-	finishReview(thread: vscode.CommentThread): Promise<void>;
-	deleteReview(): Promise<void>;
-	createOrReplyComment(thread: vscode.CommentThread): Promise<void>;
-	editComment(thread: vscode.CommentThread, comment: vscode.Comment): Promise<void>;
-	deleteComment(thread: vscode.CommentThread, comment: vscode.Comment): Promise<void>;
-}
